+++ okapi-operation-macro/src/error.rs
@@ -1,6 +1,6 @@
 use std::fmt::Display;
 
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 use quote::ToTokens;
 
 #[derive(Debug, thiserror::Error)]
@@ -9,6 +9,8 @@ pub(crate) enum Error {
     Syn(#[from] syn::Error),
     #[error(transparent)]
     Darling(#[from] darling::Error),
+    #[error("{0}")]
+    Custom(String),
 }
 
 impl Error {
@@ -20,6 +22,11 @@ impl Error {
         match self {
             Error::Syn(x) => x.into_compile_error(),
             Error::Darling(x) => x.write_errors(),
+            Error::Custom(x) => syn::Error::new(Span::call_site(), x).into_compile_error(),
         }
+    }
+
+    pub(crate) fn custom(message: impl Into<String>) -> Self {
+        Self::Custom(message.into())
     }
 }